+++ tridydb-cli/src/tests/error-unconnected-client-test-1/test.ps1
@@ -1,0 +1,3 @@
+$out = & node $PSScriptRoot/../../app.js inline --client --remote-host localhost --remote-port 0 --pretty --command "@get;" 2>&1;
+
+Write-Output $out;